Update script enqueuing order in influactive-forms plugin
Reordered the enqueueing of scripts and styles in the shortcode.php file of the influactive-forms plugin. This change allows scripts related to form submission and reCAPTCHA to load at proper points, improving form functionality and user interaction. The change maintains the readability of the code with no impact on other functionalities.