Update translation files for Influactive Forms plugin
This commit updates the translation files for the Influactive Forms WordPress plugin. Several strings in both English and French were changed to improve translation accuracy and readability. This includes strings related to form elements (such as "Add new layout" or "Delete layout"), form presentation and style (such as "Form Border width" or "Form Padding"), and error messages (such as "Bot detected" or "Failed to send email"). These updates ensure that the plugin's interface is clear and easy-to-understand for all users.