Refine Google reCAPTCHA handling in Influactive Forms
Enhanced the handling of Google reCAPTCHA in the Influactive Forms plugin. The script for Google reCAPTCHA is now enqueued in 'functions.php' instead of 'shortcode.php', only when valid site keys are available. This change reduced redundancy and streamlined the script loading. Also, a check was added to prevent the script from being enqueued multiple times.