Update line references in translation file
The line references for some strings in the translation file (POT file) for the Influactive Forms plugin have been updated. This was necessary due to recent changes in the structure of the plugin code that led to a shift in line numbers. These references are important for translators to understand where in the codebase the strings are used, to give them context for more accurate translations.