Refactor form field creation and sanitization logic
The logic for creating and sanitizing form fields in the 'edit.php' file for the 'Influactive Forms' plugin has been refactored for better readability and maintainability. New utility functions - 'sanitizeOptions', 'createField' and 'sanitizeLabel' have been added to segregate the creation and sanitization process. The 'teeny' option was removed from the array for a textarea. Changes in 'shortcode.php' reflect a simplification in the 'free_text' case - a comment was removed.