Update plugin version to 1.2
This commit updates the version of the Influactive Forms plugin, README files and all its related dependencies to 1.2. It includes updates to the PHP script and CSS files. The changes were necessary to keep consistency across all elements and improve the plugin's overall performance. It also includes a relatively newer version of the Sortable.js library for potential bug fixes or improvements.