Add Qodana code quality checking workflow
This commit adds a new GitHub workflow for Qodana code quality checking. The workflow is executed on every push event and pull requests on main branch. We utilise JetBrains Qodana action for this, which uses qodana.yaml for the configuration. This addition ensures maintaining a high standard of code quality by analyzing the code in main branch and pull requests.