Refactor form.js to improve code readability and maintainability
This commit refactors the code in 'form.js' to make it more readable and easier to maintain. Instead of using innerHTML to create elements, the code now utilizes the createElement and appendChild methods to insert new elements into the web page. This not only enhances readability but it also eliminates the potential security risks associated with the use of innerHTML.

Unnecessary code, specific to 'Sortable' was also removed as it was causing redundancy. All related changes aim to increase the efficiency and maintainability of the code without modifying its functionality.