Add Microsoft sarif eslint formatter and update package-lock.json
The commit introduces the @microsoft/eslint-formatter-sarif package as a new dependency and updates the devDependancies with eslint and eslint-config-google. This change will allow the app to generate SARIF files that can be used for static analysis in various code scanning tools. The package-lock.json file is updated to ensure that we lock the installed versions of these new dependencies.