Refactor form.js to improve code quality and readability
Code refactoring was conducted in form.js. This included changing the usage of traditional function statements to arrow functions for better readability and to maintain the current JavaScript standards. Also, script was adjusted to adhere to camelcase rule, improving overall code quality. The variable 'recaptcha_site_key' was renamed to 'recaptchaSiteKey', and function naming was introduced for xhr.onload event. ESLint directives were added to bypass some warnings. The input type declaration for 'recaptchaResponse' parameter was also updated to be more accurate.