Refactor form.js for clearer intention and consistency
In this commit, variables and their usages were refactored in 'front-end/form.js' file to enhance readability, maintainability and consistency. The type of 'messageDiv' parameter was explicitly defined as 'HTMLElement' to provide clear intention. Plus, variable 'message' was introduced to hold the 'messageDiv' value, reducing redundancy on repeating 'messageDiv.textContent'. The semicolon removed at the end of the script was a minor change to follow JavaScript's ASI (Automatic Semicolon Insertion).