Update build process & dependencies
Changes are made to streamline the build process and update project dependencies.

Converted the project to use Vite for bundling the JavaScript files. New entry points are defined in the new 'vite.config.js' file. Front-end and back-end code now get their own bundles.

'form.js' now directly imports its dependencies - 'Sortable' and 'tinymce' - instead of relying on global variables. This ensures that the dependencies are version managed and can be updated easily.

The '.gitignore' file is updated to exclude the 'dist/' folder as it contains generated code. The enqueuing scripts in 'functions.php' file are updated to load the new bundled files instead of individual files.

This refactoring enhances the maintainability of the project by using modern development tools.