Updated .eslintrc.yml to exclude dist directory
The `dist` directory has been added to the exclusion list within the .eslintrc.yml file. This ensures that ESLint, the JavaScript linter tool, will not run checks on the production-ready code. This makes sense as minified and transpiled code in the `dist` directory often don't adhere to the development code style rules. It increases efficiency by avoiding unnecessary linting on production code.