Optimize tinymce import in form.js
Removed 'tinymce' from the imports at the top of form.js file and added it to the global comment. This change is done to reduce code redundancy and to follow the convention of making global tinymce explicit in the file. This optimization may potentially result in quicker load times for the user.