Update project version and line references in language files
The project version has been updated to 'Forms everywhere by Influactive 1.2.5' in the 'influactive-forms.pot' file to reflect changes in the latest version. Additionally, line references in the same file have been corrected to synchronize with the current codebase. This detailed revision helps in maintaining accurate translation references and in ensuring the quality of localized versions.