Remove old CSS files and compile new bundled CSS
The previously separated CSS files for different components have been deleted and replaced with bundled CSS files which was achieved with the Vite bundling tool.

Due to the previous structure managing these files had become significantly more cumbersome due to the increase in complexity and size which was particularly noticeable during debugging and development processes. This change enables us to simplify the development process and the file structure of the project.