Upgrade plugin version to 1.2.6
The version of the 'Forms everywhere by Influactive' plugin is upgraded from 1.2.5 to 1.2.6. This upgrade impacts numerous files, updating the hardcoded version references to 1.2.6. The commit updates the version information in the main plugin file, scripts and styles enqueued, the plugin package.json file, translation files, and the plugin README file. This version upgrade is necessary for correctly versioning the plugin and ensuring the updated version reflects across all components, facilitating dependency management.