Add Dependabot configuration to plugin
A Dependabot configuration file (.github/dependabot.yml) was added to the plugin. This change allows the plugin dependencies to be automatically updated by Dependabot. These automatic updates help to improve the maintainability of the codebase by reducing the effort required to update dependencies. The configuration commits Dependabot to inspect for updates for npm, composer, and GitHub Actions dependencies on a daily and weekly basis.