Update ESLint version in Influactive Forms
The version of ESLint in the eslint-results.sarif file for the Influactive Forms WordPress plugin has been upgraded from 8.45.0 to 8.46.0. This change was carried out to ensure the use of the latest features and improved performance offered by the updated version. This does not affect the existing functionality of the application.