Update Influactive Plugin with metadata headers
Added metadata headers to core Influactive Forms plugin files in order to identify the package and provide licensing information. Included GPL2 license and the link to the URI, package name, and other essential details in the metadata for each respective PHP file. This ensures code maintainability and easier identification of plugins when contributing.