Improve RuntimeException information in Influactive Forms
In this commit, the exception "@throws RuntimeException If the WordPress environment is not loaded" is added to several files in the Influactive Forms plugin. This was accomplished to enhance clarity about cases when the WordPress environment is not loaded and help in error handling. Some minor changes were also done in the shortcode.php file to streamline the functionality descriptions and exception information.