Refactor return type in shortcode function
The 'influactive_form_shortcode_handler' function in the Influactive Forms plugin now returns 'void' instead of 'bool|string'. This change is made for simplification and clarity of the function's purpose, which is to generate and handle the HTML arising from the shortcode, not to return any value. As a result, developers should anticipate void return type and manage their function calls accordingly.