Improve form field security in shortcode.php
In this commit, the WordPress 'esc_attr' function has been implemented to enhance the security of a variety of form field attributes in our 'influactive-forms' plugin. This helps mitigate potential security risks that could arise from unfiltered form field values, such as potential cross-site scripting (XSS) attacks. This change ultimately safeguards the data integrity of the plugin, as well as amends the comparison operator in the 'required' field check for better validity and type checking.