Fix layout inconsistency in shortcode.php
Updated the alignment of label tags for uniformity and readability in 'shortcode.php'. Also, the use of 'esc_attr' function was increased to sanitize the output of 'label' and 'required' fields which can effectively prevent Cross-Site Scripting (XSS) attacks. The unnecessary line gap was also removed for streamlined code structure.