Corrected the code indentation in shortcode.php
This update modifies the code indentation in the 'shortcode.php' file. The file is part of the 'influactive-forms' WordPress plugin and the changes were made to improve readability and maintainability. The main logic of the plugin remains untouched. The changes consist solely of replacing spaces with tabs to follow the WordPress PHP Coding Standards.