Use sanitize_text_field for form nonce value in Influactive Forms plugin
Previously, we used the wp_unslash function to sanitize the nonce value that is being caught from the form input. This function was too general and wasn't removing certain potential harmful data types. Replaced it with sanitize_text_field function which is more secure and specifically designed for form text input sanitization ensuring a safer management of form data in the Influactive Forms plugin. Standardized all block syntaxes for exception handling as well.