Implement wp_unslash with sanitize_text_field in Influactive Forms
In this commit, we have enhanced the security of form data handling in the Influactive Forms plugin. Previously, we were using only the sanitize_text_field function on POST variables 'recaptcha_site_key' and 'recaptcha_response', which might not adequately deal with certain types of harmful data. Now, this function is being used in conjunction with wp_unslash function to ensure maximum sanitization of these values, making the plugin safer against malicious inputs.