Added check for field existence before sanitization
This change adds a check for the existence of a field in the $_POST data before performing sanitization in 'influactive-forms' plugin. Previously, calling $_POST[ $field['name'] ] directly could lead to a PHP notice if the field was not set in the post data. The new check avoids this, improving error handling and making the code more robust.