Improved the code readability with Yoda conditions and conformed to WP Coding Standards
This commit enhances the readability of the code in 'influactive-forms' plugin by implementing Yoda conditions, which is a WP recommended practice. Additionally, a safer mail sending condition has been established by using type-strict comparison with the error variable for success/failure checks. Older method of declaring an array with brackets has been replaced with the array() function to maintain WP Coding Standards. Also, a sanitization step was added to secure the field entries.