Correct variable documentation in shortcode.php
Corrected a minor error in the variable documentation of shortcode.php. The parameter $label_value was lacking a period at the end of the sentence in the comment section, which was affecting the code consistency and readability. In order to adhere to code standards and increase clarity, a period was added at the end of the comment. No functional changes were made.