Ensure secure output in captcha settings
The shortcode generating reCAPTCHA link in backend captcha settings of the InfluActive Forms plugin is refactored. Now, the public and secret site keys are escaped using the built-in WordPress function esc_attr() to prevent security issues due to unescaped user input. Moreover, the condition checking if the secret key is empty is reformatted for consistency.