Enhance variable comparisons across edit.php file
Variables comparison syntax throughout the influactive-forms plugin's edit.php file were updated as a code clean-up exercise. This crucial in preventing potential bugs that may arise due to loose comparison returns. This change does not affect the functionality of the plugin, but it establishes consistency and clarity throughout, which allows for easier debugging and understanding for future developers.