Optimized code readability in influactive-forms plugin
The adjustment to the variable comparison done in the edit.php file in the influactive-forms plugin wasn't necessarily due to functional changes. Instead, it was meant to simplify code read, provide consistency, and account for possible loose comparison returns which could lead to potential bugs in the future. Hence, it is preliminary to improve future debugging and enhance other developers' understanding.