Add phpcs.xml file and refactor text output in edit.php
Added a new phpcs.xml file to the Influactive Forms plugin for setting up PHP CodeSniffer. This will make the codebase adhere to WordPress coding standards, while excluding certain rules that may not be suitable for this project. Also, the formatting of text output in the edit.php file was refactored to improve readability. The PHP closing tag and the HTML 'option' tag were separated onto individual lines.