Refactored array declaration in 'functions.php' plugin file
Replaced the shorthand array declaration "[]" with the old style "array()" in the 'functions.php' file within the influactive-forms plugin. This change was made to maintain code consistency and to ensure compatibility with older PHP versions (<5.4) that don't support shorthand array syntax. Additionally, updated the PHPDoc annotations to accurately depict the functions' behavior.