Sanitize and refactor 'influactive-forms' plugin code
Made changes to 'influactive-forms' plugin's edit.php for better code quality, readability, and compatibility. 'esc_html' is used for key values in forms layout to prevent XSS attacks. Array syntax is changed for compatibility with PHP versions below 5.4. In code logic, replaced the direct array use to 'wp_unslash' to remove backslashes and changed the order of operations in an "if" statement for readability. Moreover, sanitized input data with suitable sanitization functions to ensure the data is safe.