Refactor Influactive form plugin for better structure
In this commit, we're restructuring the layout of the Influactive form plugin. Plugin files were moved from the root of the 'functions.php' file to their respective directories to improve code maintainability and readability. Moreover, enforced the correct naming convention for filenames by using hyphens instead of underscores. The `esc_html` function was also added in `influactive-form-fields-listing.php` file for security purposes – to ensure no harmful script is executed when outputting HTML.

This refactoring allows us to have a clearer structure of the plugin and helps in reducing potential security vulnerabilities.