Streamline data sanitization in Influactive Forms plugin
Adjusted data sanitization process in the 'influactive_form_save_post' function where an unnecessary conditional check was removed. The check was meant to handle an array conversion but it is unwarranted as 'influactive_form_fields' will always be an array. This clarifies the code, reduces its complexity and improves runtime efficiency.