Refactor field sanitization in Influactive forms plugin
This commit introduces a refactor to the input sanitization of the 'influactive_form_email_style' and 'influactive_form_email_layout' fields in the Influactive Forms plugin. Changes include usage of the 'sanitize_text_field' function. This function ensures that unnecessary or potentially harmful data is stripped out, thus improving security. Additionally, the refactoring makes the code more readable and maintainable.