Sanitize form fields directly in Influactive Forms plugin
This commit refactors the form field sanitization process of the Influactive Forms plugin. It optimizes code readibility, prevents duplication and enhances the security of the form by doing direct sanitization on each field instead of sanitizing the input 'influactive_form_fields' and then assigning it to a separate 'fields' array. This change is essential to prevent misuse of form fields as well as potential script injections.