Refactor form submission message handling in `frontForm.js`
The logic related to handling form submission messages after AJAX requests was refactored. By displaying the message directly in the `onload` and `onerror` methods of the XMLHttpRequest, we save processing time as we no longer clone `messageDiv`. The refactoring also groups the related codes more closely making it easier to understand. This makes the codebase cleaner and more efficient.