Change whitespace indentation in `backForm.js`
I replaced the variable length whitespaces used for indentation with tab characters for uniformity across the JS codebase. A proper indentation was introduced to improve the readability of the code and consistency of code formatting in back-end form handling code in 'backForm.js'. There were no changes in the underlying logic of the code, just cosmetic adjustments to adhere to the agreed coding standards for the project. Consistent indentation is important for code maintainability and readability.