Update plugin version to 1.3.2
This diff updates the version number of the Influactive Forms plugin from 1.3.1 to 1.3.2 across various files. The changes ensure consistency in the version number in the plugin's PHP, JSON, and markdown files. The version number in a PHP doc comments is also modified for consistency. The changes also correct index alignment in the comments of shortcode.php for better code readability.