Remove RuntimeException for missing form ID in Influactive Forms plugin
The RuntimeException received when form ID wasn't found has been replaced with a return statement that outputs 'false'. As this exception was causing a hard-stop in the processing of the plugin whenever form ID was not found, this change aims to avoid a potential crash of the entire WordPress system and offers a smoother way to deal with issues related to form ID missing. The Influactive Forms plugin has been updated to version 1.3.1 to reflect this change.