Update README content and tags for 'Forms Everywhere' plugin
In this commit, we updated the title, tags and description of our 'Forms Everywhere' plugin in both README.txt and README.md files to reflect updates in its features and usage. The new tag "design forms, customizable forms, email rendering, backend forms" have been added to indicate the new features being added to the plugin. The extended description now provides a brief instruction on how the plugin can be used, as well as some FAQs, enhancing clarity for potential users. This commits also updates how to install and upgrade this plugin. These changes are needed to ensure users fully understand the functionality and usage of the updated plugin.