Refactor 'Forms Everywhere' to 'Forms' for improved clarity
In this commit, 'Forms Everywhere by Influactive' was refactored to 'Forms by Influactive' in several locations for improved clarity and simplicity. The same change was applied to the names of directories and navigation instructions in both README.txt and README.md files. The decision to change this was born out of the need to reduce confusion and make it easier for users to identify and navigate the plugin in the WordPress dashboard. Mentioning 'Everywhere' felt unnecessary and potentially confusing.