Refactored dynamic form style generation code
Moved the form style generation code from `dynamic-style.php` to `shortcode.php`. The old file `dynamic-style.php` was then removed. This change is for reducing the number of files and keeping related functions together in one place for better readability and maintenance. It also involves renaming functions and actions for better consistency of the code. It also loaded form styles through admin-ajax.php, which enhances the separation between the data generation and presentation logic.