Remove unnecessary checks for Google reCaptcha
Removed checks that determine whether Google reCaptcha scripts are enqueued. This diff simplifies the function and improves performance by reducing the number of checks. The `wp_script_is` checks for 'google-captcha' and 'google-recaptcha' were unnecessary, as they are irrelevant for the strategy we use for reCaptcha handling, which is now entirely controlled from the 'influactive-forms-captcha-fields' options.