Renamed function for loading textdomain
The function load_influactive_forms_textdomain has been renamed to influactive_load_forms_textdomain to improve clarity and form a standardized naming convention. This clear naming convention ensures the function is self-explanatory and consistent with other function names in the plugin. The Add_action command was also updated to reflect this change.